web/requirements.txt
# Web Frontend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Kafka Integration
confluent-kafka>=2.3.0

# Azure ML Integration
azure-ai-ml>=1.0.0
azure-identity>=1.15.0
requests>=2.31.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
